Fix issue calculating schedule time ranges
We need to sideload teh shows and reels inorder to make the channel models schedule property accurate. If those models aren't loaded with the schedule item the end calculating erros, causing all events to be left in the collection. Ideallly we would refactor the way this works as it doesn't take advantage of data binding.